+++ metrictank/src/main/java/com/expedia/metrics/metrictank/MDMDeserializer.java
@@ -2,17 +2,16 @@ package com.expedia.metrics.metrictank;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 
 /**
  * Deserializer that reads data from Metrictank Kafka-mdm
  */
 public class MDMDeserializer {
-    private static final int METRIC_POINT_SIZE = 32;
-
+    private final MetricPointSerializer metricPointSerializer;
     private final MessagePackSerializer messagePackSerializer;
 
     public MDMDeserializer() {
+        metricPointSerializer = new MetricPointSerializer();
         messagePackSerializer = new MessagePackSerializer();
     }
 
@@ -24,26 +23,12 @@ public class MDMDeserializer {
             case 1:
                 throw new IOException("MetricDataArrayMsgp is not supported");
             case 2:
-                return new MDMData(readMetricPoint(buffer));
+                return new MDMData(metricPointSerializer.deserialize(buffer));
             case 3:
                 throw new IOException("MetricPointWithoutOrg is not supported");
             default:
                 buffer.position(buffer.position()-1);
                 return new MDMData(messagePackSerializer.deserialize(buffer));
         }
-    }
-
-    // To match format at https://github.com/raintank/schema/blob/88d6f01b3d265ddaa3c2ff110a0ed205145df57f/metricpoint.go#L56
-    private MetricPoint readMetricPoint(ByteBuffer buffer) throws IOException {
-        if (buffer.capacity() - buffer.position() != METRIC_POINT_SIZE) {
-            throw new IOException("Insufficient bytes to hold a MetricPoint");
-        }
-        final byte[] id = new byte[16];
-        buffer.order(ByteOrder.LITTLE_ENDIAN);
-        buffer.get(id);
-        final double value = buffer.getDouble();
-        final long time = Integer.toUnsignedLong(buffer.getInt());
-        final int orgId = buffer.getInt();
-        return new MetricPoint(new MetricKey(orgId, id), value, time);
     }
 }